Refactor test -> unit/test
reorganize tests to make space for integration tests
