+++ CMakeLists.txt
@@ -159,7 +159,7 @@ set(MEMORYCHECK_COMMAND_OPTIONS
 # do not enable tests if nmodl is used as submodule
 if(NOT NMODL_AS_SUBPROJECT)
   include(CTest)
-  add_subdirectory(test)
+  add_subdirectory(test/unit)
 endif()
 
 # =============================================================================
