+++ setup.cfg
@@ -1,5 +1,5 @@
 [tool:pytest]
-testpaths = test/pybind
+testpaths = test/unit/pybind
 
 [doctest]
 builder = doctest
