+++ test/unit/modtoken/modtoken.cpp
@@ -15,7 +15,7 @@
 #include "lexer/modtoken.hpp"
 #include "lexer/nmodl_lexer.hpp"
 #include "parser/nmodl_driver.hpp"
-#include "test/utils/test_utils.hpp"
+#include "test/unit/utils/test_utils.hpp"
 #include "visitors/lookup_visitor.hpp"
