+++ test/unit/parser/parser.cpp
@@ -16,8 +16,8 @@
 #include "lexer/modtoken.hpp"
 #include "parser/diffeq_driver.hpp"
 #include "parser/nmodl_driver.hpp"
-#include "test/utils/nmodl_constructs.hpp"
-#include "test/utils/test_utils.hpp"
+#include "test/unit/utils/nmodl_constructs.hpp"
+#include "test/unit/utils/test_utils.hpp"
 #include "visitors/checkparent_visitor.hpp"
 #include "visitors/lookup_visitor.hpp"
