+++ test/unit/visitor/units.cpp
@@ -12,8 +12,8 @@
 #include "ast/program.hpp"
 #include "parser/nmodl_driver.hpp"
 #include "src/config/config.h"
-#include "test/utils/nmodl_constructs.hpp"
-#include "test/utils/test_utils.hpp"
+#include "test/unit/utils/nmodl_constructs.hpp"
+#include "test/unit/utils/test_utils.hpp"
 #include "visitors/checkparent_visitor.hpp"
 #include "visitors/lookup_visitor.hpp"
 #include "visitors/units_visitor.hpp"
