+++ test/unit/visitor/var_usage.cpp
@@ -9,7 +9,7 @@
 
 #include "ast/program.hpp"
 #include "parser/nmodl_driver.hpp"
-#include "test/utils/test_utils.hpp"
+#include "test/unit/utils/test_utils.hpp"
 #include "visitors/var_usage_visitor.hpp"
 
 using namespace nmodl;
